+++ Flow/src/utopia/flow/collection/immutable/range/HasEnds.scala
@@ -1,7 +1,7 @@
 package utopia.flow.collection.immutable.range
 
 import utopia.flow.collection.immutable.Pair
-import utopia.flow.operator.{Combinable, Sign}
+import utopia.flow.operator.Sign
 import utopia.flow.operator.Sign.{Negative, Positive}
 
 import scala.math.Ordered.orderingToOrdered