+++ Flow/test/utopia/flow/test/time/WeekDayTest.scala
@@ -1,8 +1,9 @@
 package utopia.flow.test.time
 
-import utopia.flow.time.WeekDay
+import utopia.flow.time.{WeekDay, WeekDays}
 import utopia.flow.time.WeekDay.{Monday, Sunday, Thursday, Tuesday, Wednesday}
 import utopia.flow.time.TimeExtensions._
+import utopia.flow.time.WeekDays.MondayToSunday
 
 import java.time.{DayOfWeek, LocalDate}
 
@@ -13,16 +14,16 @@ import java.time.{DayOfWeek, LocalDate}
  */
 object WeekDayTest extends App
 {
+	implicit val weekdays: WeekDays = MondayToSunday
 	
-	
-	assert(WeekDay.iterate(Monday).take(7).toVector == WeekDay.values)
-	assert(WeekDay.reverseIterate(Sunday).take(7).toVector == WeekDay.values.reverse)
+	assert(WeekDay.iterate(Monday).take(7).toVector == weekdays.values)
+	assert(WeekDay.reverseIterate(Sunday).take(7).toVector == weekdays.values.reverse)
 	assert(Vector(DayOfWeek.MONDAY, DayOfWeek.TUESDAY, DayOfWeek.WEDNESDAY, DayOfWeek.THURSDAY, DayOfWeek.FRIDAY,
-		DayOfWeek.SATURDAY, DayOfWeek.SUNDAY).map { d => d: WeekDay } == WeekDay.values)
-	assert(Range.inclusive(1, 7).map(WeekDay.forIndex).toVector == WeekDay.values)
-	assert(Range.inclusive(1, 7).map { DayOfWeek.of(_): WeekDay } == WeekDay.values)
-	assert(Tuesday > Monday)
-	assert(Tuesday < Wednesday)
+		DayOfWeek.SATURDAY, DayOfWeek.SUNDAY).map { d => d: WeekDay } == weekdays.values)
+	assert(Range.inclusive(0, 6).map(weekdays.apply).toVector == weekdays.values)
+	assert(Range.inclusive(1, 7).map { DayOfWeek.of(_): WeekDay } == weekdays.values)
+	assert(weekdays.gt(Tuesday, Monday))
+	assert(weekdays.lt(Tuesday, Wednesday))
 	assert(Monday.previous == Sunday)
 	assert(Sunday.next == Monday)
 	assert(Monday.until(Tuesday) == 1.days)
@@ -32,7 +33,7 @@ object WeekDayTest extends App
 	assert(Tuesday - Thursday == 5.days)
 	
 	val today = LocalDate.now()
-	val currentWeekDay = WeekDay.current()
+	val currentWeekDay = WeekDay.current
 	
 	assert(today.next(currentWeekDay) == today + 1.weeks)
 	assert(today.next(currentWeekDay, includeSelf = true) == today)